feat: DAH-2649 separate files for modals
