+++ app/javascript/authentication/SignInForm.tsx
@@ -10,7 +10,7 @@ import {
   AlertBox,
   LinkButton,
 } from "@bloom-housing/ui-components"
-import { Link, Heading } from "@bloom-housing/ui-seeds"
+import { Dialog, Link, Heading, Alert } from "@bloom-housing/ui-seeds"
 import { FieldValues, SubmitHandler, useForm } from "react-hook-form"
 
 import { getMyAccountPath } from "../util/routeUtil"
@@ -18,12 +18,56 @@ import EmailFieldset from "../pages/account/components/EmailFieldset"
 import PasswordFieldset from "../pages/account/components/PasswordFieldset"
 import "../pages/account/styles/account.scss"
 import { AxiosError } from "axios"
+import { confirmEmail } from "../api/authApiService"
 import UserContext from "./context/UserContext"
-import {
-  AccountAlreadyConfirmedModal,
-  NewAccountNotConfirmedModal,
-} from "./components/SignInPageModals"
+import { AccountAlreadyConfirmedModal } from "./components/AccountAlreadyConfirmedModal"
 import { SiteAlert } from "../components/SiteAlert"
+
+const NewAccountNotConfirmedModal = ({
+  email,
+  onClose,
+}: {
+  email: string
+  onClose: () => void
+}) => {
+  const [emailSent, setEmailSent] = useState(false)
+  const [emailSentError, setEmailSentError] = useState<string | null>(null)
+  const requestEmail = () => {
+    confirmEmail(email)
+      .then(() => {
+        setEmailSent(true)
+      })
+      .catch(() => {
+        setEmailSentError(t("signIn.newAccount.sendEmailAgainButton.error"))
+      })
+  }
+
+  return (
+    <Dialog isOpen={!!email} onClose={onClose}>
+      <Dialog.Header>{t("signIn.newAccount.title")}</Dialog.Header>
+      {emailSent && (
+        <Alert className="sign-in-banner banner-background-color">
+          {t("signIn.newAccount.sendEmailAgainButton.confirmation")}
+        </Alert>
+      )}
+      {emailSentError && (
+        <Alert variant="alert" className="sign-in-banner">
+          {emailSentError}
+        </Alert>
+      )}
+      <Dialog.Content>{t("signIn.newAccount.p1", { email })}</Dialog.Content>
+      <Dialog.Content>{t("signIn.newAccount.p2")}</Dialog.Content>
+      <Dialog.Footer>
+        <Button type="submit" styleType={AppearanceStyleType.primary} onClick={onClose}>
+          {t("t.ok")}
+        </Button>
+        <Button styleType={AppearanceStyleType.secondary} onClick={requestEmail}>
+          {t("signIn.newAccount.sendEmailAgainButton")}
+        </Button>
+      </Dialog.Footer>
+    </Dialog>
+  )
+}
 
 const getExpiredConfirmedEmail = () => {
   const urlParams = new URLSearchParams(window.location.search)
